+++ BraintreeDropIn/BTDropInBaseViewController.m
@@ -57,20 +57,6 @@
     }];
 }
 
-- (void)showLoadingScreen:(BOOL)show animated:(BOOL)animated {
-    if (show) {
-        [self.view bringSubviewToFront:self.activityIndicatorWrapperView];
-    }
-
-    if (animated) {
-        [UIView transitionWithView:self.activityIndicatorWrapperView duration:0.2 options:UIViewAnimationOptionTransitionCrossDissolve animations:^{
-            self.activityIndicatorWrapperView.hidden = !show;
-        } completion:nil];
-    } else {
-        self.activityIndicatorWrapperView.hidden = !show;
-    }
-}
-
 - (void)showLoadingScreen:(BOOL)show {
     if (show) {
         [self.view bringSubviewToFront:self.activityIndicatorWrapperView];