+++ UnitTests/BTDropInRequestTests.m
@@ -38,11 +38,6 @@
 
     BTDropInRequest *originalRequest = [BTDropInRequest new];
     
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wdeprecated-declarations"
-    originalRequest.amount = @"10.02";
-#pragma clang diagnostic pop
-    
     originalRequest.payPalRequest = [[BTPayPalRequest alloc] initWithAmount:@"10.01"];
     originalRequest.applePayDisabled = YES;
     originalRequest.paypalDisabled = YES;
@@ -57,11 +52,6 @@
     originalRequest.allowVaultCardOverride = YES;
 
     BTDropInRequest *copiedRequest = [originalRequest copy];
-
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wdeprecated-declarations"
-    XCTAssertEqual(originalRequest.amount, copiedRequest.amount);
-#pragma clang diagnostic pop
     
     XCTAssertEqual(originalRequest.payPalRequest, copiedRequest.payPalRequest);
     XCTAssertEqual(originalRequest.applePayDisabled, copiedRequest.applePayDisabled);