+++ UnitTests/BTUIKAppearanceTests.swift
@@ -1,60 +1,6 @@
 import XCTest
 
 class BTUIKAppearanceTests: XCTestCase {
-    
-    func testLightTheme() {
-        let appearance = BTUIKAppearance.sharedInstance()
-        
-        BTUIKAppearance.lightTheme()
-        
-        XCTAssertEqual(appearance?.overlayColor, UIColor.btuik_color(fromHex:"000000", alpha:0.5))
-        XCTAssertEqual(appearance?.tintColor, UIColor.btuik_color(fromHex:"2489F6", alpha:1.0))
-        XCTAssertEqual(appearance?.disabledColor, UIColor.lightGray)
-        XCTAssertEqual(appearance?.errorForegroundColor, UIColor.btuik_color(fromHex:"ff3b30", alpha:1.0))
-        XCTAssertEqual(appearance?.switchThumbTintColor, UIColor.white)
-        XCTAssertEqual(appearance?.switchOnTintColor, UIColor.green)
-        XCTAssertEqual(appearance?.font, UIFont.systemFont(ofSize:10))
-        XCTAssertEqual(appearance?.boldFont, UIFont.boldSystemFont(ofSize:10))
-        XCTAssertEqual(appearance?.useBlurs, true)
-        XCTAssertEqual(appearance?.postalCodeFormFieldKeyboardType, .numberPad)
-        
-        XCTAssertEqual(appearance?.barBackgroundColor, UIColor.white)
-        XCTAssertEqual(appearance?.formBackgroundColor, UIColor.btuik_color(fromHex:"EFEFF4", alpha:1.0))
-        XCTAssertEqual(appearance?.formFieldBackgroundColor,  UIColor.white)
-        XCTAssertEqual(appearance?.primaryTextColor, UIColor.black)
-        XCTAssertEqual(appearance?.secondaryTextColor, UIColor.btuik_color(fromHex:"666666", alpha:1.0))
-        XCTAssertEqual(appearance?.placeholderTextColor, UIColor.lightGray)
-        XCTAssertEqual(appearance?.lineColor, UIColor.btuik_color(fromHex:"BFBFBF", alpha:1.0))
-        XCTAssertEqual(appearance?.blurStyle, .extraLight)
-        XCTAssertEqual(appearance?.activityIndicatorViewStyle, .gray)
-    }
-
-    func testDarkTheme() {
-        let appearance = BTUIKAppearance.sharedInstance()
-        
-        BTUIKAppearance.darkTheme()
-        
-        XCTAssertEqual(appearance?.overlayColor, UIColor.btuik_color(fromHex:"000000", alpha:0.5))
-        XCTAssertEqual(appearance?.tintColor, UIColor.btuik_color(fromHex:"2489F6", alpha:1.0))
-        XCTAssertEqual(appearance?.disabledColor, UIColor.lightGray)
-        XCTAssertEqual(appearance?.errorForegroundColor, UIColor.btuik_color(fromHex:"ff3b30", alpha:1.0))
-        XCTAssertEqual(appearance?.switchThumbTintColor, UIColor.white)
-        XCTAssertEqual(appearance?.switchOnTintColor, UIColor.green)
-        XCTAssertEqual(appearance?.font, UIFont.systemFont(ofSize:10))
-        XCTAssertEqual(appearance?.boldFont, UIFont.boldSystemFont(ofSize:10))
-        XCTAssertEqual(appearance?.useBlurs, true)
-        XCTAssertEqual(appearance?.postalCodeFormFieldKeyboardType, .numberPad)
-        
-        XCTAssertEqual(appearance?.barBackgroundColor, UIColor.btuik_color(fromHex:"222222", alpha:1.0))
-        XCTAssertEqual(appearance?.formBackgroundColor, UIColor.btuik_color(fromHex:"222222", alpha:1.0))
-        XCTAssertEqual(appearance?.formFieldBackgroundColor, UIColor.btuik_color(fromHex:"333333", alpha:1.0))
-        XCTAssertEqual(appearance?.primaryTextColor, UIColor.white)
-        XCTAssertEqual(appearance?.secondaryTextColor, UIColor.btuik_color(fromHex:"999999", alpha:1.0))
-        XCTAssertEqual(appearance?.placeholderTextColor, UIColor.btuik_color(fromHex:"8E8E8E", alpha:1.0))
-        XCTAssertEqual(appearance?.lineColor, UIColor.btuik_color(fromHex:"666666", alpha:1.0))
-        XCTAssertEqual(appearance?.blurStyle, .dark)
-        XCTAssertEqual(appearance?.activityIndicatorViewStyle, .white)
-    }
 
     func testLightColorScheme() {
         let appearance = BTUIKAppearance.sharedInstance()