+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Braintree iOS Drop-in SDK - Release Notes
 
-## unreleased 
+## unreleased
+
+* Remove deprecated methods and properties (NEXT_MAJOR_VERSION)
+  * Remove BTDropInBaseViewController showLoadingScreen:animated: method (NEXT_MAJOR_VERSION)
+  * Remove BTDropInRequest amount property (NEXT_MAJOR_VERSION)
+  * Remove BTUIKAppearance darkTheme and lightTheme methods (NEXT_MAJOR_VERSION)
 * Remove support for CardIO, a no longer maintained framework. (NEXT_MAJOR_VERSION)
 * Change postal code default keyboard type to `UIKeyboardTypeDefault` (NEXT_MAJOR_VERSION)
 * Update podspec to require versions `~> 4.31` of Braintree SDK dependencies.